move download button to shared
