--- src/app/components/header/queue/download-file-button/download-file-button.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DownloadFileButtonComponent } from './download-file-button.component';
-
-describe('DownloadFileButtonComponent', () => {
-  let component: DownloadFileButtonComponent;
-  let fixture: ComponentFixture<DownloadFileButtonComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ DownloadFileButtonComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(DownloadFileButtonComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
