+++ src/app/components/results-menu/scene-files/file-contents/file-contents.module.ts
@@ -10,7 +10,7 @@ import { FileNameModule } from '@components/shared/file-name';
 import { CartToggleModule } from '@components/shared/cart-toggle';
 
 import { FileContentsComponent } from './file-contents.component';
-import {DownloadFileButtonModule} from '@components/header/queue/download-file-button/download-file-button.module';
+import {DownloadFileButtonModule} from '@components/shared/download-file-button/download-file-button.module';
